pyproject.toml
[project]
name = "ha-tibber-data"
version = "1.0.0"
description = "HACS-compatible Home Assistant integration for Tibber Data API"
authors = [
    {name = "Home Assistant Community", email = "hello@home-assistant.io"}
]
readme = "README.md"
license = "MIT"
requires-python = ">=3.13"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: End Users/Desktop",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.13",
    "Topic :: Home Automation",
]
keywords = ["homeassistant", "tibber", "hacs", "integration"]
dependencies = [
    "mypy==1.18.1",
    "pytest==8.3.3",
    "pytest-asyncio==0.24.0",
    "pytest-homeassistant-custom-component==0.13.195",
    "ruff==0.13.1",
]

[project.urls]
Homepage = "https://github.com/user/ha-tibber-data"
Repository = "https://github.com/user/ha-tibber-data"
Issues = "https://github.com/user/ha-tibber-data/issues"

[build-system]
requires = ["setuptools>=77.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["custom_components*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = [
    "--strict-markers",
    "--disable-warnings",
    "--quiet"
]
filterwarnings = [
    "ignore::pytest.PytestDeprecationWarning",
    "ignore::DeprecationWarning"
]

[tool.coverage.run]
source = ["custom_components"]
omit = [
    "custom_components/tibber_data/translations/*"
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
]
